requirements.txt
faicons
shiny
shinywidgets
plotly
pandas
ridgeplot
groq
python-dotenv